requirements.txt
Django==1.8.4
django-cors-headers==1.1.0
django-extensions==1.5.2
django-pg-current-timestamp==0.2.4
django-pgjson==0.2.3
djangorestframework==3.1.1
drf-extensions==0.2.7
drf-nested-routers==0.9.0
gunicorn==19.3.0
ipdb==0.8
ipython==3.1.0
psycopg2==2.6
pudb==2015.2
git+https://github.com/dgrtwo/ParsePy.git
celery==3.1.18
django-celery==3.1.16
kombu==3.0.26
